requirements.txt
h5py
numpy
scikit-learn
pytorch